==> notebooks/01-raw-tile-processing.ipynb <==
from pathlib import Path
import geopandas
import pandas as pd

data_dir = Path("../data/")
sources_dir = data_dir / "external" / "usgs" / "tile_index"
sources = [Path(p).resolve() for p in sources_dir.glob("*.zip")]
sources

results = []
for file in sources:
    gdf = geopandas.read_file(file, ignore_geometry=True)
    results.append(
        {
            "filename": file.name,
            "first_record": gdf.iloc[0].to_dict(),
        }
    )
results

import json

source_schema_json = data_dir / "interim" / "tile_index_schema.json"
with open(source_schema_json, "w") as f:
    f.write(json.dumps(results))

from dataclasses import dataclass


@dataclass
class TileIndexSource:
    file: Path
    include_fields: list[str]
    column_rename_mapper: dict[str, str]

def create_clean_tile_index_gdf(
    tile_index_source: TileIndexSource,
) -> geopandas.GeoDataFrame:
    return geopandas.read_file(
        filename=tile_index_source.file,
        include_fields=tile_index_source.column_rename_mapper.keys(),
    ).rename(columns=tile_index_source.column_rename_mapper)

mn_sedriftless_1_2020 = TileIndexSource(
    file=data_dir
    / "external"
    / "usgs"
    / "tile_index"
    / "USGS_MN_SEDriftless_1_2021_TileIndex.zip",
    include_fields=["GridName", "WU_ID", "WU_Name"],
    column_rename_mapper={
        "GridName": "name",
        "WU_ID": "workunit_id",
        "WU_Name": "workunit_name",
    },
)
mn_sedriftless_1_2020_gdf = create_clean_tile_index_gdf(mn_sedriftless_1_2020)
mn_sedriftless_1_2020_gdf

mn_sedriftless_2_2021 = TileIndexSource(
    file=data_dir
    / "external"
    / "usgs"
    / "tile_index"
    / "USGS_MN_SEDriftless_2_2021_TileIndex.zip",
    include_fields=["GridName", "WU_ID", "WU_Name"],
    column_rename_mapper={
        "GridName": "name",
        "WU_ID": "workunit_id",
        "WU_Name": "workunit_name",
    },
)
mn_sedriftless_2_2021_gdf = create_clean_tile_index_gdf(mn_sedriftless_2_2021)
mn_sedriftless_2_2021_gdf

import json

with open(Path("../data/raw/mn_lakecounty_2018.json"), "r") as f:
    workunit_params = json.loads(f.read())
workunit_params

tile_index_dir = Path("../data/external/usgs/tile_index/")

tile_index_gdf = geopandas.read_file(
    filename=tile_index_dir / workunit_params["tile_index_source_zip"],
    include_fields=[workunit_params["tile_name_field"]],
)
tile_index_gdf["workunit_id"] = workunit_params["workunit_id"]
tile_index_gdf["workunit"] = workunit_params["workunit"]
tile_index_gdf
