==> notebooks/02-wesm-processing.ipynb <==
from pathlib import Path
import geopandas

data_dir = Path("../data/")
source = data_dir / "external" / "usgs" / "WESM.gpkg.zip"
output = data_dir / "interim" / "mn_wesm.gpkg"

filter = "workunit LIKE 'MN%' AND (ql = 'QL 1' OR ql = 'QL 0')"
gdf: geopandas.GeoDataFrame = geopandas.read_file(filename=source, where=filter)

gdf.to_file(output, driver="GPKG")

gdf

gdf.info()

gdf.at[3, "sourcedem_link"]
