==> notebooks/03-raw-tile-processing.ipynb <==
import sys

sys.path.append("..")
from src.data.build_tile_index import build_tile_index_gdf

from pathlib import Path

tile_index_dir = Path("../data/external/usgs/tile_index/").resolve()
workunit_info_dir = Path("../config/usgs_workunit_info/").resolve()

gdf = build_tile_index_gdf(tile_index_dir, workunit_info_dir)

gdf

gdf.info()

tile_index_gpkg = Path("../data/interim/tile_index.gpkg").resolve()
gdf.to_file(tile_index_gpkg, driver="GPKG", layer="tile_index")
