==> notebooks/04-sptialite-test.ipynb <==
# Configure jupysql notebook extention
%load_ext sql
%config SqlMagic.autopandas = True
%config SqlMagic.feedback = False
%config SqlMagic.displaycon = False

%sql sqlite:///:memory:

%sql SELECT load_extension('mod_spatialite')

# Failed Experiment
Loading extensions is disabled by default when sqlite is accessed in this way. I cannot figure out how to change that setting from this interface. 

Use geopandas selecting and filtering instead.

# Removed jupysql & libspatailite from environment
