==> notebooks/07-toml-config-testing.ipynb <==
import tomllib
from pathlib import Path

config_file = Path("../config/tile_index.toml").resolve()
config_file

with open(config_file, "rb") as f:
    config = tomllib.load(f)

config

config["tile_index_configs"]

from __future__ import annotations
from dataclasses import dataclass
import tomllib


@dataclass
class TileIndexConfig:
    workunit: str
    zipped_shapefile: str
    tile_name_field: str


@dataclass
class TileIndexPipelineConfig:
    tile_index_configs: list[TileIndexConfig]

    @staticmethod
    def parse_toml(config_file: Path) -> TileIndexPipelineConfig:
        with open(config_file, "rb") as f:
            return TileIndexPipelineConfig(**tomllib.load(f))

config = TileIndexPipelineConfig.parse_toml(config_file)

config

config.tile_index_configs

for tile_index_config in config.tile_index_configs:
    print(tile_index_config)
