==> notebooks/09-pystac-client-exploration.ipynb <==
## Purpose

This notebook explores using the pystac & pystac-client libraries to traverse and extract information from the ASW STAC catalog for USGS 3DEP point clouds.

from pystac_client import Client

api = Client.open("https://usgs-lidar-stac.s3-us-west-2.amazonaws.com/ept/catalog.json")

item_links = api.get_item_links()
print(f"Items in catalog: {len(item_links)}")

## Issue - STAC Catalog does not implement any helper utilities

Cannot search by name, spatial extent, content, anything..

from pystac_client import ConformanceClasses

for cc in ConformanceClasses:
    print(
        "Conforms to %s: %s"
        % (
            cc,
            api._conforms_to(cc),
        )
    )

## Implementing a href contains search

Example item relation from the top level catalog:
```javascript
{
    "rel": "item", 
    "href": "https://s3-us-west-2.amazonaws.com/usgs-lidar-stac/ept/MN_RainyLake_1_2020.json"
}
```

The workunit (`MN_RainyLake_1_2020` in this case) is used to form the item href. We can use a `str in str` conditional to filter the hrefs without calling them. This effectively prefilters the items before constructing the Item objects. 

The alternative would be to filter based on information of the Item objects returned by the `Client.get_all_items()` method. However, this will make a request to all ~2000 item urls and thus takes 2-3 minutes to run.

from pystac_client import Client
from pystac import Item


def get_ept_stac_item(stac_catalog_url: str, workunit: str) -> list[Item]:
    """Uses a `workunit in stac_catalog_url` conditional to filter item hrefs
    then constructs a list of `Item` objects that match the filter."""
    client = Client.open(stac_catalog_url)
    item_links = [link for link in client.get_item_links() if workunit in link.href]
    return [Item.from_file(item_link.href) for item_link in item_links]

stac_catalog_url = "https://usgs-lidar-stac.s3-us-west-2.amazonaws.com/ept/catalog.json"
workunit = "MN_RainyLake_1_2020"
stac_item = get_ept_stac_item(stac_catalog_url, workunit)[0]
stac_item

from pyproj import CRS

crs = CRS.from_epsg(stac_item.properties["proj:epsg"])
crs

crs.to_epsg()

epsg = stac_item.properties["proj:epsg"]
print(f"Value: {epsg} Type: {type(epsg)}")

ept_json_url = stac_item.assets["ept.json"].href
print(f"Value: {ept_json_url} Type: {type(ept_json_url)}")

## Plotting

The item is a geojson feature, so we can plot it using geopandas.

import geopandas
from geopandas import GeoDataFrame


gdf: GeoDataFrame = geopandas.read_file(stac_item.self_href, parse_dates=True)
# The gdf.explore() will raise an exception because of the datetime column
# It isn't needed for this exploration, so we just drop it
gdf.drop(columns=["datetime"], inplace=True)
gdf

gdf.filter(["id", "proj:epsg", "pc:count", "geometry"], axis="columns").explore()

## Conculsion

I'm unsure if including `pystac-client` as a dependency is worthwhile since the STAC Catalog of interest doesn't implement any of the features that would make it really helpful. I can extract the minimal information I need using the `requests` and `json` libraries instead.

Maybe there is a use in a make_ept_index intermim data process that simplifies adding tile_index data for a workunit.
