==> notebooks/11-WESM-fetch-and-filter.ipynb <==
## Fetch & Filter Workunit Extent Spatial Metadata (WESM)

**Goal**
- Fetch WESM data (.cvs or .gpkg) from hosted s3 without downloading locally
- Filter for QL 1 & QL 0 projects in Minnesota
- Write the filtered data to an interim geopackage and geoparquet

### Asset URLs

wesm_gpkg_url = (
    "https://prd-tnm.s3.amazonaws.com/StagedProducts/Elevation/metadata/WESM.gpkg"
)
wesm_csv_url = (
    "https://prd-tnm.s3.amazonaws.com/StagedProducts/Elevation/metadata/WESM.csv"
)
opr_tesm_gpkg_url = "https://prd-tnm.s3.amazonaws.com/StagedProducts/Elevation/OPR/FullExtentSpatialMetadata/OPR_TESM.gpkg"

### Download Functions

from pathlib import Path
import requests


def download_usgs_asset(url: str, output_dir: Path) -> Path:
    filename = url.split("/")[-1]
    output_file = output_dir / filename
    r = requests.get(url, stream=True)
    with open(output_file, "wb") as f:
        f.write(r.content)
    return output_file

download_usgs_asset(
    url=wesm_csv_url, output_dir=Path("../data/external/usgs/").resolve()
)

from pathlib import Path
import requests
from zipfile import ZipFile, ZIP_DEFLATED


def download_usgs_asset_to_zip(
    url: str, output_dir: Path, compression: int = ZIP_DEFLATED
) -> Path:
    filename = url.split("/")[-1]
    output_file = output_dir / f"{filename}.zip"
    r = requests.get(url, stream=True)
    with ZipFile(output_file, "w", compression=compression) as f:
        f.writestr(zinfo_or_arcname=filename, data=r.content)
    return output_file

download_usgs_asset_to_zip(
    url=wesm_csv_url, output_dir=Path("../data/external/usgs/").resolve()
)

from datetime import datetime
from email.utils import parsedate_to_datetime


def fetch_last_modified_datetime(url: str) -> datetime:
    r = requests.head(url)
    return parsedate_to_datetime(r.headers["Last-Modified"])

fetch_last_modified_datetime(wesm_csv_url)

### Filtering WESM.csv

import pandera
from pandera.typing.pandas import Series as PaSeries


class FilteredWESM(pandera.DataFrameModel):
    workunit: PaSeries[str]
    workunit_id: PaSeries[int]
    ql: PaSeries[str]
    horiz_crs: PaSeries[int]
    dem_gsd_meters: PaSeries[float]
    lpc_link: PaSeries[str]
    sourcedem_link: PaSeries[str]
    metadata_link: PaSeries[str]

    class Config:
        strict = "filter"
        coerce = True

from pathlib import Path
import pandas as pd
from pandas import DataFrame


def _filter_rows_by_workunit(df: DataFrame, workunit_contains: str) -> DataFrame:
    return df.loc[df["workunit"].str.contains(workunit_contains)]


def _filter_rows_by_quality_level(df: DataFrame, quality_level: int) -> DataFrame:
    return df.query(f"ql == 'QL {quality_level}'")


def get_filtered_wesm(
    wesm_csv: Path, workunit_contains: str = "MN_", quality_levels: list[int] = [0, 1]
) -> DataFrame:
    df = pd.read_csv(wesm_csv)
    df = _filter_rows_by_workunit(df=df, workunit_contains=workunit_contains)
    concatinated = pd.concat(
        [
            _filter_rows_by_quality_level(df=df, quality_level=ql)
            for ql in quality_levels
        ]
    )
    return FilteredWESM.validate(concatinated)

import pandas as pd

wesm_csv = download_usgs_asset(
    wesm_csv_url, output_dir=Path("../data/external/usgs/").resolve()
)

df = pd.read_csv(wesm_csv)

df.info()

mn_wesm = get_filtered_wesm(wesm_csv)
assert mn_wesm.shape[0] == 13

mn_workunits = mn_wesm["workunit"].to_list()

### Fetch EPT Catalog Items

from pystac_client import Client
from functools import partial


STAC_CATALOG_URL = "https://usgs-lidar-stac.s3-us-west-2.amazonaws.com/ept/catalog.json"


def _get_item_link(workunit: str, client: Client) -> str | None:
    item_links = [link for link in client.get_item_links() if workunit in link.href]
    return item_links[0].href if len(item_links) > 0 else None


def _add_ept_item_link_column(df: DataFrame) -> DataFrame:
    client = Client.open(STAC_CATALOG_URL)
    fetch_item_links = partial(_get_item_link, client=client)
    item_links = df["workunit"].apply(fetch_item_links)
    return df.assign(ept_item_link=item_links)

mn_wesm_with_ept = _add_ept_item_link_column(mn_wesm).dropna()

mn_wesm_with_ept["workunit_id"].to_list()

### Read tiles to geoparquet

rows = [row for row in mn_wesm_with_ept.itertuples(index=False, name="Wesm")]
rows

interim_tile_index_dir = Path("../data/interim/tile_index/").resolve()
opr_tesm_parquet = Path("../data/interim/opr_tesm.parquet").resolve()

import geopandas
from geopandas import GeoDataFrame


for row in rows:
    gdf: GeoDataFrame = geopandas.read_parquet(
        opr_tesm_parquet, filters=[("workunit_id", "=", row.workunit_id)]
    )
    output_file = interim_tile_index_dir / f"{row.workunit}.parquet"
    gdf.to_crs(epsg=row.horiz_crs).to_parquet(output_file)
    print(f"Wrote parquet for {row.workunit}")
    del gdf

parquet = Path("../data/interim/tile_index/MN_RainyLake_1_2020.parquet").resolve()
gdf: GeoDataFrame = geopandas.read_parquet(parquet)
assert gdf.crs.to_epsg() == 6344

gdf.head()
