==> notebooks/13-request-file-metadata.ipynb <==
from src.settings import Settings

settings = Settings()

url = "https://prd-tnm.s3.amazonaws.com/StagedProducts/Elevation/metadata/WESM.csv"

from pathlib import Path


old_file = Path("../data/external/usgs/WESM_old.csv").resolve()
new_file = Path("../data/external/usgs/WESM_new.csv").resolve()

from datetime import datetime
from email.utils import format_datetime

import requests

old_file_mtime = datetime.fromtimestamp(old_file.lstat().st_mtime)

old_headers = {"If-Modified-Since": format_datetime(dt=old_file_mtime)}
r_old = requests.head(url, headers=old_headers)
r_old.status_code

r_download = requests.get(url, stream=True)
with open(new_file, "wb") as f:
    f.write(r_download.content)

new_file_mtime = datetime.fromtimestamp(new_file.lstat().st_mtime)

new_headers = {"If-Modified-Since": format_datetime(dt=new_file_mtime)}
r_new = requests.head(url, headers=new_headers)
r_new.status_code
